requirements.txt
python-telegram-bot==20.0
Flask==2.0.2  # если используете Flask для webhook'ов
aiohttp
